Add MVC structure
