+++ Pr2503/Resources/Info.plist
